% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_filter.R
\name{contrast_filter}
\alias{contrast_filter}
\alias{contrast_filter_matrix}
\title{Contrast filter}
\usage{
contrast_filter(matrix, orientation = c("w", "s", "e", "n", "nw", "sw",
  "se", "ne"), thickness = 1L)

contrast_filter_matrix(orientation = c("w", "s", "e", "n", "nw", "sw",
  "se", "ne"), thickness = 1L)
}
\arguments{
\item{matrix}{The matrix input}

\item{orientation}{Which orientation should the filter have, i. e. where is
the maximum excitability? Is coded as follows: One letter stands for west,
south, east, north, two letters stand for northwest, southwest, northeast,
southeast}

\item{thickness}{How thick should the edges be? Positive integer.}
}
\description{
This function computes the contrast of a filter.
}
\section{Functions}{
\itemize{
\item \code{contrast_filter_matrix}: Provides the contrast filtering matrix.
}}

