% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-moments.R
\name{covariance}
\alias{covariance}
\title{Compute the covariance between two receptors}
\usage{
covariance(model, distance = NULL)
}
\arguments{
\item{model}{A model.}

\item{distance}{Distance between the two receptors. If NULL (default value),
returns the covariance matrix of the model. Distance may be above the number
of receptors, but assumes a circular input. The value is vectorized.}
}
\description{
This function computes the covariance between two receptors given their
distance.
}
