% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_contrast_covariance.R
\name{estimate_contrast_covariance}
\alias{estimate_contrast_covariance}
\title{Estimate covariance within a picture}
\usage{
estimate_contrast_covariance(images, orientation = NULL,
  contrast_orientation)
}
\arguments{
\item{images}{Image or images. Might be a matrix or a list of matrices}

\item{orientation}{Orientation. Can be horizontal, vertical, diagonal or
antidiagonal}

\item{contrast_orientation}{If you do not supply orientation, you may also
specify the contrast you applied and the function computes the correct
orientation.}
}
\description{
This function estimates the covariance within a picture under the assumption
that the covariance does not depend on the location of the pixel. We are
interested into the covariance along a certain dominant direction, since we
are concerned with contrast-filtered images.
}
