% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-moments.R
\name{expected_value}
\alias{expected_value}
\title{Compute the expected value}
\usage{
expected_value(model)
}
\arguments{
\item{model}{A model.}
}
\description{
This function computes the expected value of the model's input.
}
