% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dft.R
\name{frequency_decomposition}
\alias{frequency_decomposition}
\title{Frequency Decomposition}
\usage{
frequency_decomposition(process)
}
\arguments{
\item{process}{An extraclassical process}
}
\description{
Yields the frequency decomposition of an extraclassical process.
}
