% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-heatmap.R
\name{matrix_heatmap}
\alias{matrix_heatmap}
\alias{matrix_picture}
\title{Matrix Heatmap}
\usage{
matrix_heatmap(matrix)

matrix_picture(matrix)
}
\arguments{
\item{matrix}{}
}
\description{
This function generates a simple matrix heatmap using [ggplot()].
}
\section{Functions}{
\itemize{
\item \code{matrix_picture}: creates a black-white picture.
}}

