% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cifar_10.R
\name{read_cifar_10}
\alias{read_cifar_10}
\title{Read in a CIFAR-10 batch}
\usage{
read_cifar_10(batch = c("1", "2", "3", "4", "5", "test"))
}
\arguments{
\item{batch}{Which batches are read in?}
}
\description{
This function reads in batches of the CIFAR-10 dataset.
}
