% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dft.R
\name{real_dft_matrix}
\alias{real_dft_matrix}
\title{Real Discrete Fourier Transform Matrix}
\usage{
real_dft_matrix(n)
}
\arguments{
\item{n}{What dimensions should the matrix have?}
}
\description{
Instantiates a Real Discrete Fourier Transform Matrix.
}
