% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-model.R
\name{single_bar_process}
\alias{single_bar_process}
\alias{n_receptors}
\alias{length_distribution}
\alias{print.single_bar_process}
\title{Single Bar Process}
\usage{
single_bar_process(n_receptors, length_distribution)

n_receptors(model)

length_distribution(model)

\method{print}{single_bar_process}(x)
}
\arguments{
\item{n_receptors}{Number of receptors}

\item{length_distribution}{The likelihood of bars with different length. A
vector with n components where the k-th component returns the likelihood of
a bar of length k.}

\item{model}{A Model 1 type model}
}
\description{
Model 1 consists of one bar at a time that is distributed according to its
length and a set of horizontally distributed receptors.
}
